+++ src/linux.h
@@ -1,3 +1,4 @@
+//Anything specific to the linux OS should go here.
 #pragma once
 #include <cstdint>
 #include <string>
@@ -6,7 +7,7 @@
 
 
 bool DoesProcExist(pid_t pid);
-bool ReadProcMem(pid_t pid, uintptr_t address, void* buffer, size_t size);
+bool ReadProcMem(pid_t pid, uint64_t address, void* buffer, size_t size);
 
 bool ParseProcMapsLine(const std::string& line, MemoryRegion& region);
 bool ParseProcMaps(pid_t pid, std::vector<MemoryRegion>& MemRegions);
